pyproject.toml
[tool.poetry]
name = "laptopanalyzer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standart"], version = "^0.115.0"}
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
graphene-sqlalchemy = "^2.3.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
uvicorn = "^0.31.1"
python-multipart = "^0.0.12"
torch = "^2.4.1"
ultralytics = "^8.3.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
